backend/requirements.txt
Flask
Flask-Cors
gtts
moviepy
google-cloud-storage
requests
